Add configuration option for PAM autologin
This allows the user to choose whether to login automatically once PAM
succeeds for the selected user.

On one hand, automatic login makes using fingerprint login more pleasant
-- the user does not need to also press the "Log In" button, once the
fingerprint is recognized.

On the other hand, using different PAM methods, autologin might be
undesirable, as scrolling through the list of users might login a
selected user when it is undesired.

For this reason, having this functionality as a user-configurable
setting makes sense.